BLC-314 - Change the primary navigation of the demo so that it is based on a category structure.
Added a new tag-lib CategoryLookupTag.java that takes in a category name and sets a jsp attribute with the category.

Updated broadleafcommerce.tld in both locations to include CategoryLookupTag.

Changed CategoryImpl to not create a url path for a category with the url key of '/'.

Updated the load_catalog_data.sql to reflect the primary navigation for the demo site.  Also removed the test catalog and products.

Changed navigation.jsp so that the primary navigation for the demo site is built dynamically.

Fixed infinite loop bug when full url was used as category url.

Fixed a bug that wouldn't allow admin to set path for the url. For example "store/test".