Fix set sku on discrete order item so it doesn't through an exception in the case of a null retail price
In the event that you have dynamic pricing, but want to add an order item with an overridden retail price that isn't in the dynamic pricing context, the add fails because the retail retrieved is null when the order service tries to set the sku on the new item. I would expect some sort of exception to be thrown in the case that there was no retail price AND no price override, but in the case that there is a price override I would expect the price override to simply become the retail price.